Промежуточный коммит

+++ src/main/java/com/zelinskiyrk/store/product/api/request/ProductRequest.java
@@ -17,9 +17,9 @@ public class ProductRequest {
                 private ObjectId photoId;
                 private String productName;
                 private String description;
-                private Integer proteins;
-                private Integer fats;
-                private Integer carbohydrates;
-                private Integer calories;
+                private float proteins;
+                private float fats;
+                private float carbohydrates;
+                private float calories;
                 private ArrayList prices = new ArrayList();
 }
